Move npm-utils to diff package
